move template to lib/data
